pyproject.toml
[project]
name = "illu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "datasets>=3.4.1",
    "einops>=0.8.1",
    "gradio>=5.22.0",
    "moshi==0.2.2",
    "numpy==2.1.0",
    "openai>=1.66.3",
    "pydantic>=2.10.6",
    "silero-vad>=5.1.2",
    "tokenizers==0.21.1",
    "torchao==0.9.0",
    "torchaudio==2.6.0",
    "torchtune==0.5.0",
    "transformers>=4.49.0",
    "wandb>=0.19.8",
]

[dependency-groups]
dev = ["ipykernel>=6.29.5"]

[tool.setuptools]
packages = ["data_pipeline", "modeling", "train"]

[tool.uv.workspace]
members = ["modeling"]
